requirements.txt
pandas==1.2.4
python-docx==0.8.11